add, update todo and some fixes

+++ src/components/Todo/Todo.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   Card,
   Grid,
@@ -9,22 +9,19 @@ import {
   Theme,
   createStyles,
   Checkbox,
+  FormControl,
+  Input,
 } from '@material-ui/core';
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
 import { IModel } from './TodoModel';
-import { updateTodoStatus, deleteTodo } from '../../store/actions/todos';
+import { updateTodoStatus, deleteTodo, updateTodo } from '../../store/actions/todos';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
-    root: {
-      flexGrow: 1,
-      overflow: 'hidden',
-      padding: theme.spacing(0, 3),
-    },
     card: {
-      maxWidth: 575,
+      maxWidth: 675,
       margin: `${theme.spacing(1)}px auto`,
       padding: theme.spacing(2),
       transition: '0.3s',
@@ -58,10 +55,16 @@ const useStyles = makeStyles((theme: Theme) =>
       alignItems: 'center',
       justifyContent: 'flex-end',
     },
+    fullWidth: {
+      width: '100%',
+    },
   }),
 );
 
 export const Todo: React.FC<IModel> = (props) => {
+  const [rowID, setRowID] = useState<number>(-1);
+  const [todoTitle, setTodoTitle] = useState<string>('');
+
   const classes = useStyles();
   const dispatch = useDispatch();
 
@@ -70,13 +73,34 @@ export const Todo: React.FC<IModel> = (props) => {
   }
 
   const handleDeleteTodo = (id: number) => {
-    dispatch(deleteTodo(id));
+    /* eslint-disable */
+    const conf = confirm('Вы уверены? Todo будет удален!');
+
+    if (conf) {
+      dispatch(deleteTodo(id));
+    }
+  }
+
+  const handleChangeInput = (text: string) => {
+    setTodoTitle(text);
+  }
+
+  const handleUpdateTodo = (id: number) => {
+    if (todoTitle) {
+      dispatch(updateTodo(id, todoTitle));
+      setRowID(-1);
+    }
+  }
+
+  const showInput = (index: number, title: string) => {
+    setRowID(index);
+    setTodoTitle(title);
   }
 
   return (
     <Card raised={true} className={`${classes.card} ${props.todo.completed ? classes.completed : classes.notCompleted}`}>
       <Grid container spacing={2}>
-        <Grid item xs={1} className={classes.flexEnd}>
+        <Grid item lg={1} md={1} sm={2} className={classes.flexEnd}>
           <Checkbox
             checked={props.todo.completed}
             onChange={() => handleTodoComplete(props.todo.id)}
@@ -85,8 +109,25 @@ export const Todo: React.FC<IModel> = (props) => {
             inputProps={{ 'aria-label': 'secondary checkbox' }} />
         </Grid>
 
-        <Grid item xs={8}>
-          <Typography variant="h6">{props.todo.title}</Typography>
+        <Grid item lg={8} md={8} sm={10}>
+          {
+            rowID === props.index ? (
+              <FormControl className={classes.fullWidth}>
+                <Input
+                  id={`${props.todo.id}`}
+                  multiline
+                  autoFocus
+                  error={!todoTitle}
+                  onChange={(e) => handleChangeInput(e.target.value)}
+                  value={todoTitle}
+                  placeholder="Название таска"
+                  aria-describedby={`err-text-${props.todo.id}`}
+                />
+              </FormControl>
+            ) : (
+                <Typography variant="h6">{props.todo.title}</Typography>
+              )
+          }
 
           <div className={classes.flex}>
             <Typography>Исполнитель: </Typography>
@@ -96,24 +137,47 @@ export const Todo: React.FC<IModel> = (props) => {
           </div>
         </Grid>
 
-        <Grid item xs={3} className={classes.flexEnd}>
-          <ButtonGroup variant="text" color="primary" aria-label="text primary button group">
-            <Button
-              color="primary"
-              disabled={props.todo.completed}
-            >
-              edit
-            </Button>
+        <Grid item lg={3} md={3} sm={12} className={classes.flexEnd}>
+          {
+            rowID === props.index ? (
+              <ButtonGroup variant="text" color="primary" aria-label="text primary button group">
+                <Button
+                  color="default"
+                  onClick={() => setRowID(-1)}
+                  disabled={props.todo.completed}
+                >
+                  cancel
+                </Button>
 
-            <Button
-              color="secondary"
-              onClick={() => handleDeleteTodo(props.todo.id)}
-            >
-              delete
-            </Button>
-          </ButtonGroup>
+                <Button
+                  color="primary"
+                  onClick={() => handleUpdateTodo(props.todo.id)}
+                >
+                  save
+                </Button>
+              </ButtonGroup>
+            ) : (
+                <ButtonGroup variant="text" color="primary" aria-label="text primary button group">
+                  <Button
+                    color="primary"
+                    onClick={() => showInput(props.index, props.todo.title)}
+                    disabled={props.todo.completed}
+                  >
+                    edit
+                  </Button>
+
+                  <Button
+                    color="secondary"
+                    onClick={() => handleDeleteTodo(props.todo.id)}
+                  >
+                    delete
+                  </Button>
+                </ButtonGroup>
+              )
+          }
         </Grid>
       </Grid>
+
     </Card>
   );
 }
